split SharingSidebar into a few components
